+++ fast_barcode_scanner/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.0
+* Relaxed kotlin version to 1.3.50
+* Updated MLKit version (16.1.2)
+* The camera state now contains the torch state (on/off)
+* Fixed toggling the torch in the example project
+
 ## 1.0.2
 
 * Even more documentation.